pwcarmafiller.cc: import hacks/improvements to work with ATA data
This is a copy of my work which was, I think, based on a somewhat
older file, though if I look at the diff I don't think I've lost
any improvements. The copy of carmafiller in the CASA "code" SVN
has been virtually unchanged since its import. I think that I
based my work off of a file in the CARMA CVS tree, but I can't
recover the code in that tree, if it exists.