Add the MIRIAD_RECNUM column.
We record the record number of the originating MIRIAD data (specifically, the
first record for a cluster of polarization data). This is in preparation for a
gross hack to extra CASA flagging information back into MIRIAD datasets so I
can easily take advantage of the aoflagger.

In a pleasant surprise, this was actually straightforward to implement. For
once CASA doesn't let me down.